refactor/87: rename Symbol component
